requirements.txt
websockets
requests
beautifulsoup4
smolagents
litellm
python-dotenv 